requirements.txt
# Hi3DGen RunPod Worker - Production Requirements
# Python 3.10+ required
# CUDA 12.x required for GPU acceleration

# PyTorch (install with CUDA 12.4 support)
# Note: Install PyTorch separately first:
# pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# Sparse Convolution (CUDA 12.0 build)
# Note: Install separately with: pip install spconv-cu120==2.3.6
# spconv-cu120==2.3.6  # Uncomment if installing via pip

# XFormers (optimized attention)
# Note: Install separately with: pip install -U xformers --index-url https://download.pytorch.org/whl/cu124
# xformers  # Uncomment if installing via pip

# Core ML/Deep Learning
transformers==4.46.0
diffusers>=0.28.0
accelerate
pytorch-lightning==1.9.5
huggingface-hub==0.30.2
safetensors==0.4.4

# Scientific Computing
numpy==1.26.4
scipy
pandas==2.2.2
einops==0.8.0

# Computer Vision & Image Processing
opencv-python
imageio==2.36.0
scikit-image
rembg==2.0.65
# Note: realesrgan, basicsr, and gfpgan have compatibility issues with torchvision 0.21.0
# They require torchvision.transforms.functional_tensor which was removed/renamed
# These are optional for image enhancement - core pipeline works without them
# realesrgan==0.3.0  # Optional - torchvision compatibility issue
# basicsr>=1.4.2     # Optional - torchvision compatibility issue
facexlib>=0.2.5
# gfpgan>=1.3.5      # Optional - torchvision compatibility issue
Pillow

# 3D Mesh Processing
trimesh
pymeshlab==2022.2.post3
pygltflib==1.16.3
xatlas==0.0.9
open3d==0.18.0

# Configuration Management
omegaconf==2.3.0
pyyaml==6.0.2
configargparse==1.7

# Build Tools (for custom rasterizer compilation)
ninja==1.11.1.1
pybind11==2.13.4

# GPU Computing (CUDA 12.x)
cupy-cuda12x==13.4.1

# ONNX Runtime (optional, for some models)
onnxruntime==1.16.3
torchmetrics==1.6.0

# Additional ML/3D
kornia==0.8.0
timm==0.6.7
torchdiffeq

# Utilities
tqdm
psutil==6.0.0
pydantic==2.10.6

# Cloud Storage (S3-compatible)
boto3
requests

# Note: Excluded dependencies (not needed on server):
# - bpy (Blender Python API - client-side only)
# - gradio (GUI framework - not needed for serverless)
# - fastapi/uvicorn (web server - RunPod handles this)
# - PySide6 (GUI - not needed)
# - pythreejs (browser visualization - not needed)

